requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
Pillow==10.3.0
django-filter==24.2
djangorestframework-simplejwt==5.3.1